+++ CHANGELOG.md
@@ -15,6 +15,7 @@ https://github.com/nwnxee/unified/compare/build8193.36.12...HEAD
 - NoStack: Added `NWNX_NOSTACK_IGNORE_SUPERNATURAL_INNATE` to ignore effects created by the Feat, Race and SkillRanks plugins when stacking.
 - Tweaks: added `NWNX_TWEAKS_CUTSCENE_MODE_NO_TURD` to not drop a TURD when SetCutsceneMode() is called.
 - Tweaks: added `NWNX_TWEAKS_CAN_USE_ITEMS_WHILE_POLYMORPHED` to allow all items to be used while polymorphed.
+- Tweaks: added `NWNX_TWEAKS_RULESET_BARBARIAN_DR_CHANGE_FOR_PC_ONLY` to apply changes in ruleset.2da to BARBARIAN_DAMAGE_REDUCTION only to PCs. |
 
 ##### New Plugins
 - Store: Enables getting and setting store data.